fix warnings on Mac in src/gpu and include/gpu
Fix these class of warnings:
- unused functions
- unused locals
- sign mismatch
- missing function prototypes
- missing newline at end of file
- 64 to 32 bit truncation

The changes prefer to link in dead code in the debug build
with 'if (false)' than to comment it out, but trivial cases
are commented out or sometimes deleted if it appears to be
a copy/paste error.
Review URL: https://codereview.appspot.com/6296044

git-svn-id: http://skia.googlecode.com/svn/trunk@4180 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ include/gpu/GrProgramStageFactory.h
@@ -84,7 +84,7 @@ public:
         StageKey stageID = GLProgramStage::GenKey(stage);
 #if GR_DEBUG
         static const StageKey kIllegalIDMask =
-            ~((1 << kProgramStageKeyBits) - 1);
+            (uint16_t) (~((1U << kProgramStageKeyBits) - 1));
         GrAssert(!(kIllegalIDMask & stageID));
 #endif
         return fStageClassID | stageID;
